Remove afero. (#1339)
The `afero` experiment was interesting but we have discovered that the package has significant deficiencies and many ecosystem projects (e.g. https://github.com/adrg/xdg) do not currently support virtual file systems. As this is not a primary portion of what we're focusing on we don't at this time see ROI from responsibility for trying to push our corner of the ecosystem in this direction.

If we want to adopt something similar in the future we would consider https://github.com/avfs/avfs.